check for duplicate ids
previously we would only complain about duplicate ids in speficic cases
(e.g. when defining a section). now every id generated will be added
to the id hash map and an error will be generated if a collision is
detected. additionally, now the error contains a reference to the
original definition, allowing us to report to the user where the first
definition of a duplicate id is located.